requirements/base.txt
Django
psycopg2
boto
django-boto
django-crispy-forms
django-filter
django-multiselectfield
django-storages-redux
djangorestframework
Markdown
Pillow
django-extensions